pyproject.toml
[project]
name = "kyrgyz-adventure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "fastapi>=0.120.0",
    "pydentic>=0.0.1.dev3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
